multiFatoVcf_NWayFaToVcf
