+++ cmd/multiFaToVcf/multiFaToVcf_test.go
@@ -23,6 +23,7 @@ var MultiFaToVcfTests = []struct {
 	{"testdata/inputStartWithGap.fa", "chr2", "testdata/startGapTmp.vcf", "testdata/expectedStartGap.vcf", false, false, ""},
 	{"testdata/inputAltStartWithGap.fa", "chr2", "testdata/startGapAltTmp.vcf", "testdata/expectedAltStartsWithGap.vcf", false, false, ""}, //TODO: This expected file does not contain a deletion at position 1, as our parser cannot handle this case.
 	{"testdata/inputMultiSecondQueryName.fa", "chr2", "testdata/secondQueryNameTmp.vcf", "testdata/expected.vcf", false, false, "HCA"},
+	{"testdata/NWayFaToVcf/input.fa", "chr1", "testdata/NWayFaToVcf/tmp.out.vcf", "testdata/NWayFaToVcf/expected.vcf", false, false, ""},
 }
 
 func TestMultiFaVcf(t *testing.T) {
@@ -30,7 +31,7 @@ func TestMultiFaVcf(t *testing.T) {
 	for _, v := range MultiFaToVcfTests {
 		multiFaToVcf(v.inputFa, v.chromName, v.outputVcf, v.substitutionsOnlyFlag, v.retainNFlag, v.secondQueryName)
 		if !fileio.AreEqual(v.expectedVcf, v.outputVcf) {
-			t.Errorf("Error in multiFaToVcf.go, expected.vcf != output.vcf")
+			t.Errorf("Error in multiFaToVcf.go, expected.vcf != output.vcf. Expected: %v.", v.expectedVcf)
 		} else {
 			err = os.Remove(v.outputVcf)
 			exception.PanicOnErr(err)
