+++ cmd/multiFaToVcf/testdata/NWayFaToVcf/input.fa
@@ -1,0 +1,10 @@
+>ref
+AA-AAAAAAA
+>Allele1
+ACAAACACAT
+>Allele2
+AAA-ACAGAT
+>Allele3
+AAAAAAGTAT
+>Allele4
+AAAAAAAATT
